refactor(contributorbilling): move emit/ingest to internal/ [IANDT-237]
Relocate contributor billing and the entitlements ingest client under
internal/contributorbilling/ per packaging review. Add pkg/clibilling as a
thin emit/wait facade; capture and middleware remain IANDT-238.
