+++ .github/CODEOWNERS
@@ -1,1 +1,5 @@
 * @snyk/developer-experience_cli @snyk/developer-experience_ide
+
+# CLI Active Contributor billing (IANDT-237/238/240)
+/internal/contributorbilling/ @snyk/developer-experience_cli
+/pkg/clibilling/ @snyk/developer-experience_cli
