+++ internal/contributorbilling/client/entitlements_service/2026-07-29/entitlements_service.go
