+++ internal/contributorbilling/emit_integration_test.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/snyk/go-application-framework/pkg/app"
 	"github.com/snyk/go-application-framework/pkg/configuration"
-	"github.com/snyk/go-application-framework/pkg/contributorbilling"
+	"github.com/snyk/go-application-framework/internal/contributorbilling"
 )
 
 func TestEmitContributorBilling_Integration_AppliesConfigurationAndPosts(t *testing.T) {
