+++ internal/contributorbilling/emit_test.go
@@ -17,7 +17,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/snyk/go-application-framework/pkg/contributorbilling"
+	"github.com/snyk/go-application-framework/internal/contributorbilling"
 )
 
 func ingestAttributes(body map[string]interface{}) map[string]interface{} {
