+++ internal/contributorbilling/wait_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/snyk/go-application-framework/pkg/contributorbilling"
+	"github.com/snyk/go-application-framework/internal/contributorbilling"
 )
 
 func TestWait_NoPendingReturnsImmediately(t *testing.T) {
